# Prioritize edgelist representations in `to_networkx_graph`