--- a/networkx/convert.py
+++ b/networkx/convert.py
@@ -90,13 +90,6 @@ def to_networkx_graph(data, create_using=None, multigraph_input=False):
         except Exception as err:
             raise nx.NetworkXError("Input is not a correct NetworkX graph.") from err
 
-    # pygraphviz  agraph
-    if hasattr(data, "is_strict"):
-        try:
-            return nx.nx_agraph.from_agraph(data, create_using=create_using)
-        except Exception as err:
-            raise nx.NetworkXError("Input is not a correct pygraphviz graph.") from err
-
     # dict of dicts/lists
     if isinstance(data, dict):
         try:
@@ -112,6 +105,22 @@ def to_networkx_graph(data, create_using=None, multigraph_input=False):
                 return from_dict_of_lists(data, create_using=create_using)
             except Exception as err2:
                 raise TypeError("Input is not known type.") from err2
+
+    # edgelist
+    if isinstance(
+        data, list | nx.reportviews.EdgeView | nx.reportviews.EdgeDataView | Iterator
+    ):
+        try:
+            return from_edgelist(data, create_using=create_using)
+        except:
+            pass
+
+    # pygraphviz  agraph
+    if hasattr(data, "is_strict"):
+        try:
+            return nx.nx_agraph.from_agraph(data, create_using=create_using)
+        except Exception as err:
+            raise nx.NetworkXError("Input is not a correct pygraphviz graph.") from err
 
     # Pandas DataFrame
     try: